docker-web/code/requirements.txt
streamlit==1.4.0
pandas==1.3.4
urllib3==1.26.6
requests==2.26.0
openai==0.13.0
matplotlib==3.5.1
plotly==5.5.0
python-dotenv==0.19.2